Detach properly

+++ snap-agent/src/main/java/org/snapscript/agent/ConnectionChecker.java
@@ -1,16 +1,18 @@
 package org.snapscript.agent;
 
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicLong;
-
 import org.snapscript.agent.event.PingEvent;
 import org.snapscript.agent.event.PongEvent;
 import org.snapscript.agent.event.ProcessEventChannel;
 import org.snapscript.common.ThreadBuilder;
 
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicLong;
+
 public class ConnectionChecker {
-   
+
+   private final ProcessContext context;
    private final ThreadFactory factory;
    private final HealthChecker checker;
    private final AtomicBoolean active;
@@ -18,11 +20,12 @@ public class ConnectionChecker {
    private final String process;
    private final String system;
    
-   public ConnectionChecker(String process, String system) {
+   public ConnectionChecker(ProcessContext context, String process, String system) {
       this.checker = new HealthChecker(10000);
       this.factory = new ThreadBuilder();
       this.active = new AtomicBoolean();
       this.update = new AtomicLong();
+      this.context = context;
       this.process = process;
       this.system = system;
    }
@@ -36,12 +39,27 @@ public class ConnectionChecker {
          .build();
       
       try {
+         CountDownLatch latch = context.getLatch();
+         ProcessMode mode = context.getMode();
          long time = System.currentTimeMillis();
-         
-         if(!channel.send(pong)) {
-            ProcessTerminator.terminate("Ping failed for " + process);
+
+         if(mode.isDetachRequired()) {
+            long count = latch.getCount();
+
+            if(count > 0) { // send pong only if still running
+               if(!channel.send(pong)) {
+                  ProcessTerminator.terminate("Ping failed for " + process);
+               } else {
+                  update.set(time);
+               }
+            }
+         } else {
+            if(!channel.send(pong)) {
+               ProcessTerminator.terminate("Ping failed for " + process);
+            } else {
+               update.set(time);
+            }
          }
-         update.set(time);
       } catch(Exception e) {
          e.printStackTrace();
          ProcessTerminator.terminate("Ping failed for " + process + " with " + e);
@@ -65,6 +83,8 @@ public class ConnectionChecker {
       
       @Override
       public void run() {
+         ProcessMode mode = context.getMode();
+
          try {
             while(true) {
                Thread.sleep(frequency);
@@ -79,7 +99,9 @@ public class ConnectionChecker {
          } catch(Exception e) {
             e.printStackTrace();
          } finally {
-            ProcessTerminator.terminate("Connection checker timeout elapsed");
+            if(mode.isTerminateRequired()) {
+               ProcessTerminator.terminate("Connection checker timeout elapsed");
+            }
          }
       }
    }
